Add base test scene for timeline
